test(sidecar): the solar tests follow the modules they exercise
tests/test_solar.py was named for a file that no longer exists and covered both
sides of the split: 24 tests on the service and 19 on the product, with seven
helpers between them.

test_sun.py takes the POWER reader, the record it returns and the position of
the sun over it. test_pv.py takes the array chain, what the terrain does to the
irradiance reaching it, the season windows and the colour policy.

The two tests that read both sides are on the product side, because
terra.energy depends on terra.sun and not the other way round. Of the seven
helpers only _three_years is used by both, and it is one call to
pd.date_range: it is written in both files rather than reached for across
them, which for two lines is the smaller cost.

Test count is unchanged at 43. Four imports the split left idle are dropped.

Verification: pytest 511 passed, 8 skipped.
